+++ playground/test.play.ts
@@ -1,7 +1,4 @@
+import * as path from 'path'
+import * as url from 'url'
 
-function hello(...args) {
-  console.log(args)
-}
-
-const aaa = []
-hello(...aaa)
+console.log(url.resolve('http://bynaki.iptime.org:1330', 'graphql'))